requirements.txt
dj-database-url==2.3.0
Django==5.1.4
django-debug-toolbar==4.4.6
django-ninja==1.3.0
django-phonenumber-field[phonenumberslite]==8.0.0
environs==11.2.1
Pillow==11.0.0
